Add 'Investors_reaction' notebook to calculate and visualise investor reactions
This commit introduces the 'Investors_reaction' notebook, which calculates
growth and logarithmic returns based on normalised prices for five
companies discussed in our study.

Additionally, the notebook constructs and analyses an equally weighted
portfolio index of these stocks, calculating both growth and log returns
to estimate investor reactions.

Visualisations are included for both individual company performances
and the overall index.

Due to copyright restrictions, stock price data and the
'SPI_daily.csv' file are not included in the repository.